requirements.txt
boto3==1.14.41
botocore==1.17.41
docutils==0.15.2
future==0.18.2
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
